backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
qrcode[pil]>=7.4
reportlab>=4.0.0
requests>=2.31.0
Pillow>=10.0.0

